refactor: extract process::stream_and_wait for child output plumbing
run.rs and clone.rs both spawned two tokio tasks to read child stdout
and stderr line-by-line. Extract the shared machinery into a utility,
parameterized on whether stderr should be labelled (shell commands) or
passed through (git progress).
