+++ src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod path;
+pub mod process;
 pub mod shell;
 pub mod sudo;
